+++ event-tracking/event-tracking-discovery/src/main/java/uk/gov/justice/eventsourcing/discovery/workers/EventDiscoveryWorker.java
@@ -1,23 +1,83 @@
 package uk.gov.justice.eventsourcing.discovery.workers;
 
-import java.util.List;
-import javax.inject.Inject;
+import static java.lang.String.format;
+
+import uk.gov.justice.eventsourcing.discovery.dataaccess.EventSubscriptionStatus;
+import uk.gov.justice.eventsourcing.discovery.dataaccess.EventSubscriptionStatusRepository;
+import uk.gov.justice.services.common.util.UtcClock;
+import uk.gov.justice.services.event.buffer.core.repository.subscription.NewStreamStatusRepository;
+import uk.gov.justice.services.eventsourcing.discovery.DiscoveryResult;
+import uk.gov.justice.services.eventsourcing.discovery.EventSubscriptionDiscoveryBean;
+import uk.gov.justice.services.eventsourcing.repository.jdbc.discovery.StreamPosition;
 import uk.gov.justice.subscription.SourceComponentPair;
-import uk.gov.justice.subscription.SubscriptionSourceComponentFinder;
+
+import java.util.Optional;
+import java.util.UUID;
+
+import javax.inject.Inject;
+
+import org.slf4j.Logger;
 
 public class EventDiscoveryWorker {
 
     @Inject
-    private SubscriptionSourceComponentFinder subscriptionSourceComponentFinder;
+    private EventSubscriptionStatusRepository eventSubscriptionStatusRepository;
 
     @Inject
-    private EventDiscoverer eventDiscoverer;
+    private NewStreamStatusRepository newStreamStatusRepository;
 
-    public void runEventDiscovery() {
+    @Inject
+    private EventSubscriptionDiscoveryBean eventSubscriptionDiscoveryBean;
 
-        final List<SourceComponentPair> sourceComponentPairs = subscriptionSourceComponentFinder
-                .findListenerOrIndexerPairs();
+    @Inject
+    private Logger logger;
 
-        sourceComponentPairs.forEach(eventDiscoverer::runEventDiscoveryForSourceComponentPair);
+    @Inject
+    private UtcClock clock;
+
+    public void runEventDiscoveryForSourceComponentPair(final SourceComponentPair sourceComponentPair) {
+
+        final String source = sourceComponentPair.source();
+        final String component = sourceComponentPair.component();
+        final Optional<EventSubscriptionStatus> eventSubscriptionStatusOptional = eventSubscriptionStatusRepository.findBy(
+                source,
+                component);
+
+        if (eventSubscriptionStatusOptional.isPresent()) {
+            logger.debug(format("Running event discovery for source '%s' component '%s'", source, component));
+            final EventSubscriptionStatus eventSubscriptionStatus = eventSubscriptionStatusOptional.get();
+
+            final Optional<UUID> latestKnownEventId = eventSubscriptionStatus.latestEventId();
+
+            final DiscoveryResult discoveryResult = eventSubscriptionDiscoveryBean.discoverNewEvents(latestKnownEventId);
+
+            discoveryResult.streamPositions()
+                    .forEach(streamPosition -> runDiscoveryFor(streamPosition, source, component));
+
+            if (discoveryResult.latestKnownEventId().isPresent()) {
+                eventSubscriptionStatusRepository.save(new EventSubscriptionStatus(
+                        source,
+                        component,
+                        discoveryResult.latestKnownEventId(),
+                        clock.now()
+                ));
+            }
+        }
+    }
+
+    private void runDiscoveryFor(final StreamPosition streamPosition, final String source, final String component) {
+
+        final UUID streamId = streamPosition.streamId();
+        final Long latestKnownPosition = streamPosition.positionInStream();
+
+        logger.debug(format("Updating latest known position to '%d' for stream id '%s', source '%s' component '%s'", latestKnownPosition, streamId, source, component));
+
+        newStreamStatusRepository.upsertLatestKnownPosition(
+                streamId,
+                source,
+                component,
+                latestKnownPosition,
+                clock.now()
+        );
     }
 }